Add common RolesAndGroupsBaseDeleteDialog
ref #3967
